+++ x/concentrated-liquidity/pool_test.go
@@ -784,7 +784,7 @@ func (s *KeeperTestSuite) TestGetUserUnbondingPositions() {
 // - Ensure that the pool accumulator is updates
 // - Ensure that the position accumulators are updated
 // - Ensures that the position 1 receives incentives  but not position 2
-func (s *KeeperTestSuite) TestMigrateAccumulatorToScalingFactor() {
+func (s *KeeperTestSuite) TestMigrateIncentivesAccumulatorToScalingFactor() {
 	const incentiveDenom = "uosmo"
 
 	var emissionRatePerSecDec = osmomath.OneDec()
@@ -850,7 +850,7 @@ func (s *KeeperTestSuite) TestMigrateAccumulatorToScalingFactor() {
 	s.Require().NoError(err)
 
 	// System under test.
-	err = s.App.ConcentratedLiquidityKeeper.MigrateAccumulatorToScalingFactor(s.Ctx, poolID)
+	err = s.App.ConcentratedLiquidityKeeper.MigrateIncentivesAccumulatorToScalingFactor(s.Ctx, poolID)
 	s.Require().NoError(err)
 
 	// Ensure that the pool accumulator has been properly migrated
@@ -903,6 +903,92 @@ func (s *KeeperTestSuite) TestMigrateAccumulatorToScalingFactor() {
 
 	// Ensure that position 2 cannot claim any incentives
 	s.validateClaimableIncentives(positionTwoID, sdk.NewCoins())
+}
+
+func (s *KeeperTestSuite) TestMigrateSpreadFactorAccumulatorToScalingFactor() {
+	s.SetupTest()
+	s.App.ConcentratedLiquidityKeeper.SetSpreadFactorPoolIDMigrationThreshold(s.Ctx, 1000)
+
+	spreadRewardAccumValue := sdk.NewDecCoins(sdk.NewDecCoinFromDec(USDC, sdk.MustNewDecFromStr("276701288297")))
+	positionAccumValue := sdk.NewDecCoins(sdk.NewDecCoinFromDec(USDC, sdk.MustNewDecFromStr("276701288297").Quo(sdk.MustNewDecFromStr("2"))))
+
+	// Create CL pool that will not be migrated
+	concentratedPool := s.PrepareCustomConcentratedPool(s.TestAccs[0], ETH, USDC, DefaultTickSpacing, osmomath.MustNewDecFromStr("0.003"))
+	poolIDNonMigrated := concentratedPool.GetId()
+
+	// Create CL pool that will be migrated
+	concentratedPool = s.PrepareCustomConcentratedPool(s.TestAccs[0], ETH, USDC, DefaultTickSpacing, osmomath.MustNewDecFromStr("0.003"))
+	poolIDMigrated := concentratedPool.GetId()
+
+	// Create a position in pool that will not be migrated
+	poolNonMigrated, err := s.App.ConcentratedLiquidityKeeper.GetPoolById(s.Ctx, poolIDNonMigrated)
+	s.Require().NoError(err)
+	poolNonMigratedPositionID, _ := s.CreateFullRangePosition(poolNonMigrated, DefaultCoins)
+
+	// Create a position in pool that will be migrated
+	poolMigrated, err := s.App.ConcentratedLiquidityKeeper.GetPoolById(s.Ctx, poolIDMigrated)
+	s.Require().NoError(err)
+	poolMigratedPositionID, _ := s.CreateFullRangePosition(poolMigrated, DefaultCoins)
+
+	// Manually set spread reward accumulator for pool that will not be migrated
+	feeAccumulatorNonMigrated, err := s.App.ConcentratedLiquidityKeeper.GetSpreadRewardAccumulator(s.Ctx, poolIDNonMigrated)
+	s.Require().NoError(err)
+	feeAccumulatorNonMigrated.AddToAccumulator(spreadRewardAccumValue)
+
+	// Manually set spread reward accumulator for position that will not be migrated
+	nonMigratedPositionAccumulatorKey := types.KeySpreadRewardPositionAccumulator(poolNonMigratedPositionID)
+	feeAccumulatorNonMigrated.SetPositionIntervalAccumulation(nonMigratedPositionAccumulatorKey, positionAccumValue)
+
+	// Manually set spread reward accumulator for pool that will be migrated
+	feeAccumulatorMigrated, err := s.App.ConcentratedLiquidityKeeper.GetSpreadRewardAccumulator(s.Ctx, poolIDMigrated)
+	s.Require().NoError(err)
+	feeAccumulatorMigrated.AddToAccumulator(spreadRewardAccumValue)
+
+	// Manually set spread reward accumulator for position that will be migrated
+	migratedPositionAccumulatorKey := types.KeySpreadRewardPositionAccumulator(poolMigratedPositionID)
+	feeAccumulatorMigrated.SetPositionIntervalAccumulation(migratedPositionAccumulatorKey, positionAccumValue)
+
+	// Non-migrated pool claim
+	nonMigratedPoolBeforeUpgradeSpreadFactor, err := s.App.ConcentratedLiquidityKeeper.GetClaimableSpreadRewards(s.Ctx, poolNonMigratedPositionID)
+	s.Require().NoError(err)
+	s.Require().NotEmpty(nonMigratedPoolBeforeUpgradeSpreadFactor)
+
+	// Migrated pool claim
+	migratedPoolBeforeUpgradeSpreadFactor, err := s.App.ConcentratedLiquidityKeeper.GetClaimableSpreadRewards(s.Ctx, poolMigratedPositionID)
+	s.Require().NoError(err)
+	s.Require().NotEmpty(migratedPoolBeforeUpgradeSpreadFactor)
+
+	// System under test.
+	err = s.App.ConcentratedLiquidityKeeper.MigrateSpreadFactorAccumulatorToScalingFactor(s.Ctx, poolIDMigrated)
+	s.Require().NoError(err)
+
+	// Manually change the pool IDs list to the pool ID in the test
+	types.MigratedSpreadFactorAccumulatorPoolIDsV25 = map[uint64]struct{}{}
+	types.MigratedSpreadFactorAccumulatorPoolIDsV25[poolIDMigrated] = struct{}{}
+
+	// Non-migrated pool: ensure that the claimable spread rewards are the same before and after migration
+	nonMigratedPoolAfterUpgradeSpreadFactor, err := s.App.ConcentratedLiquidityKeeper.GetClaimableSpreadRewards(s.Ctx, poolNonMigratedPositionID)
+	s.Require().NoError(err)
+	s.Require().Equal(nonMigratedPoolBeforeUpgradeSpreadFactor.String(), nonMigratedPoolAfterUpgradeSpreadFactor.String())
+
+	// Migrated pool: ensure that the claimable spread rewards are the same before and after migration
+	migratedPoolAfterUpgradeSpreadFactor, err := s.App.ConcentratedLiquidityKeeper.GetClaimableSpreadRewards(s.Ctx, poolMigratedPositionID)
+	s.Require().NoError(err)
+	s.Require().Equal(migratedPoolBeforeUpgradeSpreadFactor.String(), migratedPoolAfterUpgradeSpreadFactor.String())
+
+	// Position's accumulator for non migrated pool should not be updated
+	feeAccumulatorNonMigrated, err = s.App.ConcentratedLiquidityKeeper.GetSpreadRewardAccumulator(s.Ctx, poolIDNonMigrated)
+	s.Require().NoError(err)
+	nonMigratedPositionAfterMigration, err := feeAccumulatorNonMigrated.GetPosition(nonMigratedPositionAccumulatorKey)
+	s.Require().NoError(err)
+	s.Require().Equal(positionAccumValue.String(), nonMigratedPositionAfterMigration.AccumValuePerShare.String())
+
+	// Position's accumulator for migrated pool should be updated
+	feeAccumulatorMigrated, err = s.App.ConcentratedLiquidityKeeper.GetSpreadRewardAccumulator(s.Ctx, poolIDMigrated)
+	s.Require().NoError(err)
+	migratedPositionAfterMigration, err := feeAccumulatorMigrated.GetPosition(migratedPositionAccumulatorKey)
+	s.Require().NoError(err)
+	s.Require().Equal(positionAccumValue.MulDecTruncate(cl.PerUnitLiqScalingFactor).String(), migratedPositionAfterMigration.AccumValuePerShare.String())
 }
 
 // Basic test to validate that positions are correctly returned for a pool